requirements.txt
pandas
numpy
scikit-learn
transformers
torch
streamlit
langchain_groq
python-dotenv
chromadb  
pymupdf  
sentence-transformers
